Add Phone UI tests and simplify the overall testing code.
