+++ ElectricSidecar/UICatalog/Assets.xcassets/AccentColor.colorset/Contents.json
