+++ ElectricSidecar/UICatalog/Assets.xcassets/AppIcon.appiconset/Contents.json
