+++ ElectricSidecar/UICatalog/Assets.xcassets/Contents.json
