+++ ElectricSidecar/UICatalogTests/UITestCase.swift
@@ -1,21 +1,12 @@
-import WatchKit
 import XCTest
 
-private func sanitizedDeviceName() -> String {
-  return WKInterfaceDevice.current().name
-    .replacing(/Clone \d+ of\ /, with: "")
-    .trimmingCharacters(in: .whitespaces)
-}
+class UITestCase: XCTestCase {
 
-final class WidgetUITests: XCTestCase {
-
-  var testEnvironment: String = "Undefined"
   var snapshotDirectory: String!
   var app: XCUIApplication!
   override func setUpWithError() throws {
     continueAfterFailure = true
 
-    testEnvironment = sanitizedDeviceName()
     snapshotDirectory = ProcessInfo.processInfo.environment["SNAPSHOT_PATH"] ?? ""
     if snapshotDirectory.isEmpty {
       snapshotDirectory = ProcessInfo.processInfo.environment["TMPDIR"]
@@ -30,31 +21,7 @@ final class WidgetUITests: XCTestCase {
     app.terminate()
   }
 
-  func testVehicleCharge() throws {
-    app.launchEnvironment = [
-      "TESTING": "1",
-      "test-case": "vehicle-charge-widget"
-    ]
-    app.launch()
-
-    let rootView = app.otherElements.containing(.any, identifier: "root-view").firstMatch
-    XCTAssertTrue(rootView.exists)
-    try writeSnapshot(sanitizedSnapshot(rootView.screenshot().image))
-  }
-
-  func testVehicleRange() throws {
-    app.launchEnvironment = [
-      "TESTING": "1",
-      "test-case": "vehicle-range-widget"
-    ]
-    app.launch()
-
-    let rootView = app.otherElements.containing(.any, identifier: "root-view").firstMatch
-    XCTAssertTrue(rootView.exists)
-    try writeSnapshot(sanitizedSnapshot(rootView.screenshot().image))
-  }
-
-  private func writeSnapshot(
+  func writeSnapshot(
     _ image: UIImage,
     named name: String? = nil,
     file: StaticString = #file,
@@ -94,25 +61,6 @@ final class WidgetUITests: XCTestCase {
     return string
       .replacingOccurrences(of: "\\W+", with: "-", options: .regularExpression)
       .replacingOccurrences(of: "^-|-$", with: "", options: .regularExpression)
-  }
-
-  private func sanitizedSnapshot(_ image: UIImage) throws -> UIImage {
-    UIGraphicsBeginImageContextWithOptions(image.size, true, image.scale)
-    image.draw(at: CGPoint.zero)
-    guard let context = UIGraphicsGetCurrentContext() else {
-      throw FailedToGetCurrentGraphicsContext()
-    }
-    let statusBarHeight = app.statusBars.firstMatch.frame.height
-    let path = UIBezierPath(rect: CGRect(x: image.size.width - 70, y: 0, width: 70, height: statusBarHeight))
-    context.addPath(path.cgPath)
-    context.clip()
-    UIColor.black.setFill()
-    context.fill(CGRect(x: 0, y: 0, width: image.size.width, height: image.size.height))
-    guard let newImage = UIGraphicsGetImageFromCurrentImageContext() else {
-      throw FailedToGetGetImageFromCurrentContext()
-    }
-    UIGraphicsEndImageContext()
-    return newImage
   }
 }
