+++ ElectricSidecar/Widgets/VehicleCharge/VehicleChargeWidget.swift
@@ -93,7 +93,7 @@ private struct WidgetView : View {
 #endif
 
     default:
-      Text("Unsupported")
+      Text("\(family.debugDescription)")
     }
   }
